fix: locate `sentry-cli` when working with a mono-repo
The current approach to locating the `sentry-cli` executable makes the assumption that the package is installed in an Ember app at the root of the repository. This works fine most of the time, but breaks in a mono-repo where the actual `node_modules` directory might be above the current project in the directory structure.

This change dynamically locates the `sentry-cli` executable by:

1. Dynamically locating the `package.json` for the `@sentry/cli` package in a way that will work for both standard and mono-repo setups
2. Extracting the root of the `@sentry/cli` package from that path, which should be safe; the `package.json` will always be at the package root
3. Reading the `package.json` file for `@sentry/cli` to determine where the executable actually lives
4. Building an absolute path to the `sentry-cli` executable from the information that we've gathered

This makes the approach to locating `sentry-cli` resilient to changes in the directory structure for different types of apps