feat: allow for configurable `releaseName`
This package currently generates a name for the release in Sentry that is not configurable by the user. If they want to use a release name that differs from the convention hard-coded into the package -- like, for example, using just the revision number -- that is currently impossible (aside from doing something to modify your dependencies).

This adds a `releaseName` configuration option that can be used to override the default name provided by the package. If left un-configured, the previous pattern is used. However, it can also be used to provide a different value that will be used to name the release in Sentry instead.